03/05/2024 Fixed geo stories icons hiding. Fixed large intersect counter. Msg buttons now work. Geo story publishing now handles answers. Added changing profile pictures (now changed instantly). Added clicked friend overlay

+++ app/src/main/java/tagme/CustomIconOverlay.kt
@@ -51,7 +51,7 @@ class CustomIconOverlay(
 
             val distance = sqrt((x - it.x).toDouble().pow(2.0) + (y - it.y).toDouble().pow(2.0))
 
-            val iconRadius = (drawable.intrinsicWidth * overlayScale / 2).toFloat()
+            val iconRadius = (drawable.intrinsicWidth * overlayScale / 2)
 
             if (distance <= iconRadius) {
                 clickListener?.invoke(this)
